requirements.txt
# Install: pip install Flask==3.0.0
Flask==3.0.0

# Install: pip install flask-cors==4.0.0
flask-cors==4.0.0

# Install: pip install Flask-Bcrypt==1.0.1
Flask-Bcrypt==1.0.1

# Install: pip install Werkzeug==3.0.1
Werkzeug==3.0.1

# Install: pip install google-generativeai==0.5.4
google-generativeai==0.5.4

# Install: pip install docformatter==1.7.5
docformatter==1.7.5

# Install: pip install pydocstyle==6.3.0
pydocstyle==6.3.0

# (Optional) If using .env file
# Install: pip install python-dotenv==1.0.1
python-dotenv==1.0.1
